+++ test/support/helpers.rb
@@ -2,7 +2,7 @@ class Minitest::Test
   include ActiveJob::TestHelper
 
   def setup
-    [Product, Store].each do |model|
+    [Product, Store, Contact].each do |model|
       setup_model(model)
     end
   end
@@ -29,6 +29,7 @@ class Minitest::Test
     Searchkick.callbacks(:bulk) do
       model.destroy_all
     end
+    model.searchkick_index.reindex_queue.clear if Searchkick.redis
   end
 
   def store(documents, model = default_model, reindex: true)